requirements.txt
Requests==2.31.0
pypdfium2
torch
numpy
pandas
tqdm
nltk
marker-pdf  
sentence-transformers
qdrant-client
langchain-community
